Finished implementation of social media alt redirects
Also cleaned up the javascript side of whoogle config so that it now
uses arrays of available fields for parsing config values instead of manually assigning each
one to a variable.

This doesn't include support for Google Maps > Open Street Maps, that
seems a bit more involved than the social media redirects were, so it
should likely be a separate effort.