Override __str__ for endpoint f-strings
By default, f-strings use __str__ not __repr__, unless supplied with a
"!r" at the end of the str. Rather than going through and making the
strings work for the overridden __repr__ for endpoints, it's easier just
to convert the method over to __str__.

Also cleaned up some broken routes and formatting.