Expand `/window` endpoint to behave like a proxy
The current `/window` endpoint is used as a proxy currently, but only
for removing Javascript from the result page. This expands the existing
functionality to allow users to proxy search result pages through their
Whoogle instance.

This commit temporarily overwrites the nojs feature until I can decide
on how I want to approach separating (or not separating) these two
features.

Note that this feature will likely be merged in without being 100%
finished, since I'm unsure of how users intend to use the feature. With
some community testing, I'll get a better idea of A) how many people are
actually going to use this feature, and B) what aspects of the feature
are useful or not useful.