Force using Lynx UA for all searches
This is a possible fix for the recent removal of javascript-free search
results by Google. Forcing the Lynx user agent to be used for all
searches returns search results, although they're formatted a bit
differently than before. If this proves to be a long term fix, styling
can be fixed at a later date.

See #1211